Update TickTickProjectData model to change SortOrder type from int to long for improved compatibility with TickTick API

+++ backend/SecondBrain.Api/Models/Integrations/TickTickProjectData.cs
@@ -28,6 +28,6 @@ namespace SecondBrain.Api.Models.Integrations
         public string Name { get; set; } = string.Empty;
 
         [JsonPropertyName("sortOrder")]
-        public int SortOrder { get; set; }
+        public long SortOrder { get; set; }
     }
 }
